+++ src/lib/djerba/plugins/supplement/body/plugin.py
@@ -4,6 +4,7 @@ import os
 from djerba.plugins.base import plugin_base, DjerbaPluginError
 import djerba.core.constants as core_constants
 from djerba.util.render_mako import mako_renderer
+import djerba.util.assays as assays
 import djerba.util.input_params_tools as input_params_tools
 
 class main(plugin_base):
@@ -23,32 +24,42 @@ class main(plugin_base):
         self.set_ini_default(core_constants.ATTRIBUTES, 'clinical')
         self.set_ini_default(self.FAILED, "False")
         self.set_priority_defaults(self.DEFAULT_CONFIG_PRIORITY)
-    
+
+    def check_assay_name(self, wrapper):
+        [ok, msg] = assays.name_status(wrapper.get_my_string(self.ASSAY))
+        if not ok:
+            self.logger.error(msg)
+            raise DjerbaPluginError(msg)
+
     def configure(self, config):
         config = self.apply_defaults(config)
         wrapper = self.get_config_wrapper(config)
         # Get input_data.json if it exists; else return None
         input_data = input_params_tools.get_input_params_json(self)
         if input_data == None:
-            msg = "Input_params.json does not exist. Parameters must be set manually."
-            self.logger.warning(msg)
-
+            msg = "File input_params.json does not exist. Parameters must be set manually."
+            self.logger.info(msg)
         if wrapper.my_param_is_null(self.ASSAY):
-            wrapper.set_my_param(self.ASSAY, input_data[self.ASSAY])
-
-        return config
+            if input_data:
+                wrapper.set_my_param(self.ASSAY, input_data[self.ASSAY])
+            else:
+                msg = "Cannot find assay from input_params.json or manual config"
+                self.logger.error(msg)
+                raise DjerbaPluginError(msg)
+        self.check_assay_name(wrapper)
+        return wrapper.get_config()
 
     def extract(self, config):
         wrapper = self.get_config_wrapper(config)
-        #TO DO: add a check that assay type is a permitted value
+        self.check_assay_name(wrapper)
         data = {
             'plugin_name': self.identifier+' plugin',
             'priorities': wrapper.get_my_priorities(),
             'attributes': wrapper.get_my_attributes(),
             'merge_inputs': {},
             'results': {
-                'assay': config[self.identifier][self.ASSAY],
-                'failed': config[self.identifier][self.FAILED]
+                'assay': wrapper.get_my_string(self.ASSAY),
+                'failed': wrapper.get_my_boolean(self.FAILED)
             },
             'version': str(self.SUPPLEMENT_DJERBA_VERSION)
         }